+++ R/create_deathvar.R
@@ -23,7 +23,7 @@
 #'   with blank space " ". Default is ULORSAK.
 #' @param valsclass optional argument specifying the class of the
 #'   comorbidities/outcomes. Allowed values are "num" (numeric: 1, 0),
-#'   "char" (character: "yes", "no"), "fac" (factor: "yes", "no").
+#'   "char" (character: "Yes", "No"), "fac" (factor: "Yes", "No").
 #'   Default is "num".
 #' @param meta_reg Optional argument specifying registries used. Printed in
 #'   metadatatable. Default is
@@ -116,16 +116,13 @@ create_deathvar <- function(cohortdata,
     }
   }
 
-  if (valsclass %in% c("char", "fac")) {
+  if (valsclass == "char") {
     out_data <- out_data %>%
-      mutate(!!name2 := case_when(
-        !!sym(name2) == 1 ~ "Yes",
-        TRUE ~ "No"
-      ))
-    if (valsclass == "fac") {
-      out_data <- out_data %>%
-        mutate(!!name2 := factor(!!sym(name2)))
-    }
+      mutate(!!name2 := if_else(!!sym(name2) == 1, "Yes", "No"))
+  }
+  if (valsclass == "fac") {
+    out_data <- out_data %>%
+      mutate(!!name2 := factor(!!sym(name2), levels = c(0, 1), labels = c("No", "Yes")))
   }
 
   # create meta data to print in table in statistical report